# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateStoreRequest {
  storeName: string;
}

export type StoreDataPayloadTopUpRates = {[key: string]: number};

export type StoreDataPayloadGarmentRateOverrides = {[key: string]: { [key: string]: unknown }};

export type StoreDataPayloadStoreInfo = {[key: string]: string};

export interface StoreDataPayload {
  storeName: string;
  customers: unknown[];
  orders: unknown[];
  nextDI: number;
  topUpRates: StoreDataPayloadTopUpRates;
  garmentRateOverrides: StoreDataPayloadGarmentRateOverrides;
  storeInfo: StoreDataPayloadStoreInfo;
}

export type StoreDataResponse = StoreDataPayload & {
  storeId: string;
  updatedAt: string;
};

export interface ErrorResponse {
  error: string;
}