# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stores
    description: Multi-device store data sync for DRYCU-72H
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stores:
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store
      description: Creates a new store with a freshly generated shareable code and empty data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "200":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDataResponse"
  /stores/{storeId}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Fetch a store's synced data
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDataResponse"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: pushStore
      tags: [stores]
      summary: Push (upsert) a store's full data snapshot
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreDataPayload"
      responses:
        "200":
          description: Store data saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDataResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateStoreRequest:
      type: object
      properties:
        storeName:
          type: string
      required:
        - storeName
    StoreDataPayload:
      type: object
      properties:
        storeName:
          type: string
        customers:
          type: array
          items: {}
        orders:
          type: array
          items: {}
        nextDI:
          type: integer
        topUpRates:
          type: object
          additionalProperties:
            type: number
        garmentRateOverrides:
          type: object
          additionalProperties:
            type: object
        storeInfo:
          type: object
          additionalProperties:
            type: string
      required:
        - storeName
        - customers
        - orders
        - nextDI
        - topUpRates
        - garmentRateOverrides
        - storeInfo
    StoreDataResponse:
      allOf:
        - $ref: "#/components/schemas/StoreDataPayload"
        - type: object
          properties:
            storeId:
              type: string
            updatedAt:
              type: string
          required:
            - storeId
            - updatedAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error