# lib/api-zod/src/generated/types/storeDataPayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreDataPayloadGarmentRateOverrides } from './storeDataPayloadGarmentRateOverrides';
import type { StoreDataPayloadStoreInfo } from './storeDataPayloadStoreInfo';
import type { StoreDataPayloadTopUpRates } from './storeDataPayloadTopUpRates';

export interface StoreDataPayload {
  storeName: string;
  customers: unknown[];
  orders: unknown[];
  nextDI: number;
  topUpRates: StoreDataPayloadTopUpRates;
  garmentRateOverrides: StoreDataPayloadGarmentRateOverrides;
  storeInfo: StoreDataPayloadStoreInfo;
}