# lib/api-zod/src/generated/types/storeDataResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreDataPayload } from './storeDataPayload';

export type StoreDataResponse = StoreDataPayload & {
  storeId: string;
  updatedAt: string;
};